+++ carbon-usage-client/src/components/consumption/consumption-list/ConsumptionList.tsx
@@ -1,13 +1,12 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { Consumption, ConsumptionPatchPayload } from '@/utils/types';
 import { useUser } from '@/contexts/UserContext';
-import { ACTIVITY_TYPES } from '@/utils/constants';
 import FilterSection from './FilterSection';
 import PaginationControls from './PaginationControls';
 import ConsumptionTable from './ConsumptionTable';
 import { useConsumptionData } from '@/hooks/useConsumptionData';
+import { useEditingState } from '@/hooks/useEditingState';
 
 export default function ConsumptionList() {
   const navigate = useNavigate();
@@ -37,20 +36,17 @@ export default function ConsumptionList() {
 
   const { userContext } = useUser();
 
-  // Edit States
-  const [editingId, setEditingId] = useState<number | null>(null);
-  const [editForm, setEditForm] = useState<{
-    activity_type_table_id: number;
-    amount: number;
-    date: string;
-  } | null>(null);
-  const [changedFields, setChangedFields] = useState<ConsumptionPatchPayload>(
-    {}
-  );
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
+  const {
+    editingId,
+    editForm,
+    isSubmitting,
+    handleEdit,
+    handleCancelEdit,
+    handleInputChange,
+    handleSaveEdit,
+    handleDelete,
+  } = useEditingState(fetchConsumptions, userContext);
   // Sorting
-
   const [showFilters, setShowFilters] = useState(false);
 
   useEffect(() => {
@@ -100,139 +96,6 @@ export default function ConsumptionList() {
       month: 'short',
       day: 'numeric',
     });
-  };
-
-  const handleDelete = async (id: number) => {
-    if (!window.confirm('Are you sure you want to delete this consumption?')) {
-      return;
-    }
-
-    try {
-      const response = await fetch(
-        `http://localhost:3000/consumption/delete/${id}`,
-        {
-          method: 'DELETE',
-          credentials: 'include',
-        }
-      );
-
-      if (!response.ok) {
-        throw new Error(`Error: ${response.status}`);
-      }
-
-      if (userContext && userContext.userId) {
-        await fetchConsumptions({
-          userId: userContext.userId,
-        });
-      }
-    } catch (err) {
-      console.error('Failed to delete consumption', err);
-      alert('Failed to delete consumption');
-    }
-  };
-
-  const handleEdit = (consumption: Consumption) => {
-    const matchingActivity = ACTIVITY_TYPES.find(
-      (activity) => activity.name === consumption.activity_table.name
-    );
-
-    const activityId = matchingActivity
-      ? matchingActivity.activity_type_id
-      : consumption.activity_type_table_id;
-
-    setEditingId(consumption.id);
-    setEditForm({
-      activity_type_table_id: activityId,
-      amount: consumption.amount,
-      date: consumption.date.split('T')[0],
-    });
-    setChangedFields({});
-  };
-
-  const handleCancelEdit = () => {
-    setEditingId(null);
-    setEditForm(null);
-    setChangedFields({});
-  };
-
-  const handleInputChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>
-  ) => {
-    if (!editForm) return;
-
-    const { name, value } = e.target;
-
-    const updatedForm = { ...editForm };
-
-    if (name === 'activity_type_table_id') {
-      updatedForm.activity_type_table_id = parseInt(value, 10);
-
-      const selectedActivity = ACTIVITY_TYPES.find(
-        (activity) => activity.activity_type_id === parseInt(value, 10)
-      );
-
-      if (selectedActivity) {
-        setChangedFields((prev) => ({
-          ...prev,
-          activity_type_table_id: parseInt(value, 10),
-          activity_type_name: selectedActivity.name,
-          emission_factor: selectedActivity.emission_factor,
-          unit: selectedActivity.unit,
-        }));
-      }
-    } else if (name === 'amount') {
-      updatedForm.amount = parseFloat(value);
-      setChangedFields((prev) => ({
-        ...prev,
-        amount: parseFloat(value),
-      }));
-    } else if (name === 'date') {
-      updatedForm.date = value;
-      setChangedFields((prev) => ({
-        ...prev,
-        date: value,
-      }));
-    }
-
-    setEditForm(updatedForm);
-  };
-
-  const handleSaveEdit = async () => {
-    if (!editingId || !editForm || Object.keys(changedFields).length === 0) {
-      handleCancelEdit();
-      return;
-    }
-
-    try {
-      setIsSubmitting(true);
-
-      const response = await fetch(
-        `http://localhost:3000/consumption/patch/${editingId}`,
-        {
-          method: 'PATCH',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(changedFields),
-          credentials: 'include',
-        }
-      );
-
-      if (!response.ok) {
-        throw new Error(`Error: ${response.status}`);
-      }
-
-      if (userContext && userContext.userId) {
-        await fetchConsumptions({
-          userId: userContext.userId,
-        });
-      }
-
-      handleCancelEdit();
-    } catch (err) {
-      console.error('Failed to update consumption', err);
-      alert('Failed to update consumption');
-    } finally {
-      setIsSubmitting(false);
-    }
   };
 
   const toggleFilters = () => {
